migration_service/requirements.txt
flask==2.3.3
google-cloud-storage==2.10.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
requests==2.31.0
